fix: ensure JSON serializability of all model outputs (fixes #141)
- Added json_convert utility in sklearn_migrator/utils.py to recursively convert NumPy ndarrays and scalars to native Python types.
- Centralized version_tuple helper in utils.py to reduce duplication.
- Applied json_convert to all serialize_* functions across classification, regression, clustering, and dimension modules.
- Verified fix with reproduction script and existing test suite.
