+++ sklearn_migrator/classification/gradient_boosting_clf.py
@@ -4,10 +4,9 @@ from sklearn.ensemble import GradientBoostingClassifier
 from ..regression.decision_tree_reg import serialize_decision_tree_reg
 from ..regression.decision_tree_reg import deserialize_decision_tree_reg
 from sklearn.dummy import DummyClassifier
+from ..utils import json_convert, version_tuple
 
 import sklearn
-
-version_sklearn = sklearn.__version__
 
 
 all_features = [
@@ -39,35 +38,6 @@ all_features = [
     'n_trees_per_iteration_',
     'presort'
 ]
-
-
-def version_tuple(version: str) -> tuple:
-    """
-    Convert a version string into a comparable tuple of integers.
-
-    Parameters
-    ----------
-    version : str
-        Version string (e.g. '1.2.0').
-
-    Returns
-    -------
-    tuple
-        Tuple of integers (major, minor, patch).
-    """
-
-    version_split = version.split('.')
-
-    if len(version_split) == 1:
-        new_version = (int(version_split[0]), 0, 0)
-    elif len(version_split) == 2:
-        new_version = (int(version_split[0]), int(version_split[1]), 0)
-    elif len(version_split) == 3:
-        new_version = (int(version_split[0]), int(version_split[1]), int(version_split[2]))
-    else:
-        new_version = 'Formato no valido'
-
-    return new_version
 
 
 if version_tuple(sklearn.__version__) < version_tuple('1.4.0'):
@@ -200,7 +170,7 @@ def serialize_gradient_boosting_clf(model: GradientBoostingClassifier, version_i
     metadata['other_params'] = other_params
     metadata['version_sklearn_in'] = version_in
 
-    return metadata
+    return json_convert(metadata)
 
 def deserialize_gradient_boosting_clf(data: dict, version_out: str) -> GradientBoostingClassifier:
     """
