+++ sklearn_migrator/classification/knn_clf.py
@@ -1,5 +1,6 @@
 import numpy as np
 from sklearn.neighbors import KNeighborsClassifier
+from ..utils import json_convert
 
 all_features = [
     "_fit_X",
@@ -53,7 +54,7 @@ def serialize_knn_clf(model: KNeighborsClassifier, version_in: str) -> dict:
     metadata["other_params"] = other_params
     metadata["version_sklearn_in"] = version_in
 
-    return metadata
+    return json_convert(metadata)
 
 def deserialize_knn_clf(data: dict, version_out: str) -> KNeighborsClassifier:
     """
