+++ sklearn_migrator/regression/mlp_reg.py
@@ -1,6 +1,7 @@
 import warnings
 import numpy as np
 from sklearn.neural_network import MLPRegressor
+from ..utils import json_convert
 
 all_features = [
     'batch_size',
@@ -104,7 +105,7 @@ def serialize_mlp_reg(model: MLPRegressor, version_in: str) -> dict:
     metadata['other_params'] = other_params
     metadata['version_sklearn_in'] = version_in
 
-    return metadata
+    return json_convert(metadata)
 
 
 def deserialize_mlp_reg(data: dict, version_out: str) -> MLPRegressor:
